footerを追加、いちぶ修正
